add AbstractTemporaryRector
